context: non-passive loopback links also active a driver
If a driver node is linked with a non-passive link to itself, it
should also become running.

One example of this is the Midi bridge, that needs to start when you
link one device to another. The devices are all on the same node
so linking devices would result in loopback connections.