context: add method to get a work_queue
Make a method to get a work-queue from the context. There is one
work-queue to use and the context will allocate it when requested
and free when destroyed.

The work queue is handy to delay execution of some logic for later,
either in the next iteration of the main-loop or when an async
operation completed.

Export some work-queue methods.